ref 同步跳转处理, 回调通知下沉到各通道实现中

+++ daxpay-multi/daxpay-multi-service-controller/src/main/java/cn/daxpay/multi/service/controller/order/PayOrderController.java
@@ -23,6 +23,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
+
 /**
  * 支付订单控制器
  * @author xxm
@@ -68,7 +70,7 @@ public class PayOrderController {
     @RequestPath("查询金额汇总")
     @Operation(summary = "查询金额汇总")
     @GetMapping("/getTotalAmount")
-    public Result<Integer> getTotalAmount(PayOrderQuery param){
+    public Result<BigDecimal> getTotalAmount(PayOrderQuery param){
         return Res.ok(queryService.getTotalAmount(param));
     }
